global: refactor current_theme_icons
* current_theme_icons doesn't need the application context anymore
